# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "M1 Local Journal",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/unlock": {
      "post": {
        "summary": "Unlock",
        "operationId": "unlock_api_v1_auth_unlock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Unlock"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/entries": {
      "post": {
        "summary": "Create",
        "operationId": "create_api_v1_entries_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Create"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Entries",
        "operationId": "entries_api_v1_entries_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 1024
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entries/{entry_id}": {
      "get": {
        "summary": "Get",
        "operationId": "get_api_v1_entries__entry_id__get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Edit",
        "operationId": "edit_api_v1_entries__entry_id__patch",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Patch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete",
        "operationId": "delete_api_v1_entries__entry_id__delete",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Delete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entries/{entry_id}/revisions": {
      "get": {
        "summary": "Revisions",
        "operationId": "revisions_api_v1_entries__entry_id__revisions_get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports/preview": {
      "post": {
        "summary": "Preview",
        "operationId": "preview_api_v1_exports_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Selector"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports": {
      "post": {
        "summary": "Export",
        "operationId": "export_api_v1_exports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Export"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/invitations": {
      "post": {
        "summary": "Invitation",
        "operationId": "invitation_api_v1_sync_invitations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpochReset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sync/devices/{device_id}/revoke": {
      "post": {
        "summary": "Revoke",
        "operationId": "revoke_api_v1_sync_devices__device_id__revoke_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpochReset"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/epoch": {
      "post": {
        "summary": "Reset Epoch",
        "operationId": "reset_epoch_api_v1_sync_epoch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpochReset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/device/pair": {
      "post": {
        "summary": "Pair",
        "operationId": "pair_api_v1_device_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pair"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/device/operations": {
      "post": {
        "summary": "Apply",
        "operationId": "apply_api_v1_device_operations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Packet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Create": {
        "properties": {
          "operation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operation Id"
          },
          "entry_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entry Id"
          },
          "base_revision": {
            "type": "integer",
            "const": 0,
            "title": "Base Revision"
          },
          "payload": {
            "$ref": "#/components/schemas/EntryInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "entry_id",
          "base_revision",
          "payload"
        ],
        "title": "Create"
      },
      "Delete": {
        "properties": {
          "operation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operation Id"
          },
          "base_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Base Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "base_revision"
        ],
        "title": "Delete"
      },
...
      "EntryPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntryOutput"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "EntryPage"
      },
...
      "EpochReset": {
        "properties": {
          "prune_tombstones": {
            "type": "boolean",
            "title": "Prune Tombstones",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EpochReset"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Patch": {
        "properties": {
          "operation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operation Id"
          },
          "base_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Base Revision"
          },
          "changes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Changes"
          },
          "confirm_type_change": {
            "type": "boolean",
            "title": "Confirm Type Change",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "base_revision",
          "changes"
        ],
        "title": "Patch"
      },
      "Receipt": {
        "properties": {
          "entry_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entry Id"
          },
          "operation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operation Id"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "result_code": {
            "type": "string",
            "enum": [
              "MAC_SAVED",
              "DELETED"
            ],
            "title": "Result Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "entry_id",
          "operation_id",
          "revision",
          "result_code"
        ],
        "title": "Receipt"
      },
      "RevisionPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntryRevisionOutput"
            },
            "type": "array",
            "title": "Items"
          },
          "next_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next After"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "next_after"
        ],
        "title": "RevisionPage"
      },
      "Selector": {
        "properties": {
          "ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array",
                "maxItems": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Ids"
          },
          "types": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "inbox",
                    "daily",
                    "sleep",
                    "creative"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Types"
          },
          "date_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date From"
          },
          "date_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date To"
          },
          "include_history": {
            "type": "boolean",
            "title": "Include History",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Selector"
      },
      "Unlock": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "Unlock"
      },
...
    }
  }
}